mastering_bitcoin/requirements.txt
python-bitcoinrpc
python-bitcoinlib